requirements.txt
appdirs==1.4.4
attrs==19.3.0
bcrypt==3.2.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudant==2.13.0
CouchDB==1.2
cryptography==3.4.7
fastapi==0.60.1
h11==0.9.0
httpcore==0.12.3
httptools==0.1.1
httpx==0.16.1
idna==2.10
passlib==1.7.4
pathspec==0.8.0
pycparser==2.20
pycrypto==2.6.1
pydantic==1.6.1
PyJWT==2.0.1
python-dotenv==0.15.0
python-multipart==0.0.5
regex==2020.7.14
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
uvicorn==0.11.6
uvloop==0.14.0
websockets==8.1